requirements.txt
pandas 
fastapi 
strawberry-graphql 
uvicorn 